Add basic TUV-x bindings to Julia interface
- add tuvs to the build
- modularize the julia bindings to split micm/tuvx
- report tuvx version number
- Dropping Julia 1.10 since its bundled libcurl lacks
  the CURL_OPENSSL_4 version node that the distribution HDF5 needs
