+++ .gitignore
@@ -48,3 +48,5 @@ julia/Manifest.toml
 julia/deps/lib/
 
 .claude
+build-jll/
+build-julia/
