+++ cmake/dependencies.cmake
@@ -188,7 +188,7 @@ if (MUSICA_ENABLE_TUVX AND MUSICA_BUILD_C_CXX_INTERFACE AND NOT MUSICA_USE_PREBU
   # NOTE: `docker/Dockerfile.tuvx` extracts TUVX_GIT_REPOSITORY and TUVX_GIT_TAG
   #       from this script to set up tests against stand-alone TUV-x
   set_git_default(TUVX_GIT_REPOSITORY https://github.com/NCAR/tuv-x.git)
-  set_git_default(TUVX_GIT_TAG v0.16.0)
+  set_git_default(TUVX_GIT_TAG 496a07c2dcd8ccea8787b9bbee24cc6a20f4bf7f)  # NCAR/tuv-x#217, merged: adds the from host radiative transfer solver
 
   FetchContent_Declare(tuvx
     GIT_REPOSITORY ${TUVX_GIT_REPOSITORY}
