+++ julia/Project.toml
@@ -13,7 +13,7 @@ YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
 CxxWrap = "0.17"
 JSON = "0.21, 1"
 YAML = "0.4"
-julia = "1.10 - 1.12"
+julia = "1.11 - 1.12"
 Musica_jll = ">= 0.16"
 
 [extras]
